Add safer concurrency for small in-process caches: Done (mutex added earlier). Implement one prioritized follow-up: Done (made debug counters safe in non-atomic builds by adding a mutex; atomics remain used when available). Run static analysis / linters and fix findings: Partially done — ran stricter compiler warnings and fixed many issues; cppcheck wasn't installed so I used the compiler as the static checker and fixed high-confidence warnings. Rebuild & smoke test: Done. What I changed (concrete)
olsrd_status_plugin.c
Added static pthread_mutex_t g_kv_cache_lock (previous change) and used it to guard the small in-process caches (cache_get/cache_set, olsr_cache_get/olsr_cache_set).
Added static pthread_mutex_t g_debug_lock and updated non-atomic debug macros to lock/unlock around updates/loading (safe fallback when C11 atomics are unavailable).
Fixed a misplaced #endif that was introduced during editing.
Updated APP2 macro to cast int lengths to size_t to avoid signedness arithmetic problems.
connections.c
Updated APPEND / JAPPEND macros to use (size_t) casts for appended lengths, and to do length arithmetic using size_t.

+++ src/connections.c
@@ -81,14 +81,14 @@ int render_connections_plain(char **buf_out, size_t *len_out){
     /* snprintf returns the number of characters that would have been written; cap copy to buffer size-1 */ \
     if (_app_n >= (int)sizeof(t)) _app_n = (int)sizeof(t) - 1; \
     if (_app_n > 0) { \
-      if (len + _app_n + 1 > cap) { \
-        while (cap < len + _app_n + 1) cap *= 2; \
+      if (len + (size_t)_app_n + 1 > cap) { \
+        while (cap < len + (size_t)_app_n + 1) cap *= 2; \
         char *nb2 = (char*)realloc(buf, cap); \
         if (!nb2) { free(buf); return -1; } \
         buf = nb2; \
       } \
       memcpy(buf + len, t, (size_t)_app_n); \
-      len += _app_n; \
+      len += (size_t)_app_n; \
       buf[len] = 0; \
     } \
   } while(0)
@@ -160,14 +160,14 @@ int render_connections_json(char **buf_out, size_t *len_out){
     int _app_n = snprintf(t, sizeof(t), fmt, ##__VA_ARGS__); \
     if (_app_n >= (int)sizeof(t)) _app_n = (int)sizeof(t) - 1; \
     if (_app_n > 0) { \
-      if (len + _app_n + 1 > cap) { \
-        while (cap < len + _app_n + 1) cap *= 2; \
+      if (len + (size_t)_app_n + 1 > cap) { \
+        while (cap < len + (size_t)_app_n + 1) cap *= 2; \
         char *nb2 = (char*)realloc(buf, cap); \
         if (!nb2) { free(buf); return -1; } \
         buf = nb2; \
       } \
       memcpy(buf + len, t, (size_t)_app_n); \
-      len += _app_n; \
+      len += (size_t)_app_n; \
       buf[len] = 0; \
     } \
   } while(0)
